db scrub: declare the fixture functions' id column as bigint
`Test (Docker)` finally completed a run and reported 48 passed, 34 failed.
Thirty-two of those are the `prosqlbody` regression the previous two
commits fix. The other two are these fixtures, which had never executed
before — they were added in the last few rounds and every Docker run since
was cancelled by the next push.

Both failed before reaching a single assertion:

    a_view_read_through_a_tracked_function_is_refreshed_in_order
      syntax error at or near "BEGIN"

    a_view_read_through_an_untracked_function_is_refused
      return type mismatch in function declared to return record
      Final statement returns bigint instead of integer at column 1
      (SQL function "bridge_fn")

The first is `BEGIN ATOMIC` on the module's default `postgres:11-alpine`,
already fixed one commit ago by pinning those two tests to `16-alpine`.

The second is independent of the server version and is fixed here:
`SAMPLE_SCHEMA` declares `users.id` as `BIGSERIAL`, while every one of
these fixture functions declared `RETURNS TABLE(id int, email text)`.
PostgreSQL rejects that at CREATE FUNCTION time on every version. I hit
exactly this error in a scratch fixture while developing the feature and
corrected it there, and never carried the correction into the test file —
which nothing caught, because these tests had not run.

Six declarations, all `int` -> `bigint`.

Re-measured against a schema matching `SAMPLE_SCHEMA`, running each
fixture's SQL verbatim and then the command:

    fn_opaque       -> refused `a_report via bridge_fn`,   users 200
    fn_via_view     -> refused `a_report via bridge_fn`,   users 200
    fn_nested_opaque-> refused `a_report via opaque_inner`, users 200
    fn_atomic       -> scrubs; z_source refreshed FIRST, both views clean

which is what each test asserts. The two fixtures added in the previous
commit were checked the same way rather than assumed:

    operator, tracked body -> scrubs; z_source FIRST, 0 seeded addresses
    operator, opaque body  -> refused `a_report via opaque_op`,  users 200
    domain,   opaque body  -> refused `a_report via opaque_dom`, users 200

+++ autumn-cli/tests/integration/db_scrub.rs
@@ -2477,7 +2477,7 @@ async fn a_view_read_through_an_untracked_function_is_refused() {
     opaque
         .batch_execute(
             "CREATE MATERIALIZED VIEW z_source AS SELECT id, email FROM users; \
-             CREATE FUNCTION bridge_fn() RETURNS TABLE(id int, email text) \
+             CREATE FUNCTION bridge_fn() RETURNS TABLE(id bigint, email text) \
                  AS $$ SELECT id, email FROM z_source $$ LANGUAGE sql; \
              CREATE MATERIALIZED VIEW a_report AS SELECT * FROM bridge_fn();",
         )
@@ -2509,7 +2509,7 @@ async fn a_view_read_through_an_untracked_function_is_refused() {
     indirect
         .batch_execute(
             "CREATE MATERIALIZED VIEW z_source AS SELECT id, email FROM users; \
-             CREATE FUNCTION bridge_fn() RETURNS TABLE(id int, email text) \
+             CREATE FUNCTION bridge_fn() RETURNS TABLE(id bigint, email text) \
                  AS $$ SELECT id, email FROM z_source $$ LANGUAGE sql; \
              CREATE VIEW bridge_view AS SELECT * FROM bridge_fn(); \
              CREATE MATERIALIZED VIEW a_report AS SELECT * FROM bridge_view;",
@@ -2537,9 +2537,9 @@ async fn a_view_read_through_an_untracked_function_is_refused() {
     nested
         .batch_execute(
             "CREATE MATERIALIZED VIEW z_source AS SELECT id, email FROM users; \
-             CREATE FUNCTION opaque_inner() RETURNS TABLE(id int, email text) \
+             CREATE FUNCTION opaque_inner() RETURNS TABLE(id bigint, email text) \
                  AS $$ SELECT id, email FROM z_source $$ LANGUAGE sql; \
-             CREATE FUNCTION outer_fn() RETURNS TABLE(id int, email text) LANGUAGE sql \
+             CREATE FUNCTION outer_fn() RETURNS TABLE(id bigint, email text) LANGUAGE sql \
                  BEGIN ATOMIC SELECT f.id, f.email FROM opaque_inner() f \
                      WHERE EXISTS (SELECT 1 FROM countries); END; \
              CREATE MATERIALIZED VIEW a_report AS SELECT * FROM outer_fn();",
@@ -2590,9 +2590,9 @@ async fn a_view_read_through_a_tracked_function_is_refreshed_in_order() {
     atomic
         .batch_execute(
             "CREATE MATERIALIZED VIEW z_source AS SELECT id, email FROM users; \
-             CREATE FUNCTION atomic_fn() RETURNS TABLE(id int, email text) LANGUAGE sql \
+             CREATE FUNCTION atomic_fn() RETURNS TABLE(id bigint, email text) LANGUAGE sql \
                  BEGIN ATOMIC SELECT id, email FROM z_source; END; \
-             CREATE FUNCTION atomic_outer() RETURNS TABLE(id int, email text) LANGUAGE sql \
+             CREATE FUNCTION atomic_outer() RETURNS TABLE(id bigint, email text) LANGUAGE sql \
                  BEGIN ATOMIC SELECT id, email FROM atomic_fn(); END; \
              CREATE MATERIALIZED VIEW a_report AS SELECT * FROM atomic_outer();",
         )
